# Nodejs Best Api Practices Solution/swagger.yaml
openapi: 3.0.0
info:
  title: RESTful API Example
  version: 1.0.0
paths:
  /api/v1/campaigns:
    get:
      summary: List campaigns
      responses:
        '200':
          description: OK
    post:
      summary: Create a campaign
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Campaign'
      responses:
        '201':
          description: Created
  /api/v1/campaigns/{id}:
    get:
      summary: Get a campaign
      parameters:
        - in: path
          name: id
          schema:
            type: integer
      responses:
        '200':
          description: OK
    put:
      summary: Update a campaign
      parameters:
        - in: path
          name: id
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Campaign'
      responses:
        '200':
          description: Updated
    delete:
      summary: Delete a campaign
      parameters:
        - in: path
          name: id
          schema:
            type: integer
      responses:
        '204':
          description: No Content
  /api/v1/donations:
    post:
      summary: Create a donation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Donation'
      responses:
        '201':
          description: Created

components:
  schemas:
    Campaign:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        goal:
          type: number
        end_date:
          type: string
          format: date
        email:
          type: string
    Donation:
      type: object
      properties:
        campaignId:
          type: integer
        amount:
          type: number
        donorName:
          type: string
        donorEmail:
          type: string
        cardNumber:
          type: string
        expirationDate:
          type: string
        cvv:
          type: string